+++ ruby/hokusai/mounting/update_entry.rb
@@ -34,7 +34,7 @@ module Hokusai
               if !!visible
                 if child.else_condition_active?
                   meta.child_delete(index) if child_present.call(child, false)
-                  child.else_active = 0
+                  child.else_active = false
                 end
 
                 unless child_present.call(child, true)
@@ -55,8 +55,8 @@ module Hokusai
                   meta.children!.insert(index, child_block)
 
                   child_block.send(:before_updated) if child_block.respond_to?(:before_updated)
-                  child_block.update
-                  child.else_active = 0
+                  Hokusai.update(child_block)
+                  child.else_active = false
                 end
               elsif !visible
                 if !child.has_else_condition? || (child.has_else_condition? && !child.else_condition_active?)
@@ -82,10 +82,10 @@ module Hokusai
                   child_block = NodeMounter.new(node, else_child_block_klass, [stack], previous_providers: providers).mount(context: context, providers: providers)
                   UpdateEntry.new(child_block, block, utarget).register(context: context, providers: providers)
                   meta.children!.insert(index, child_block)
-
                   child_block.send(:before_updated) if child_block.respond_to?(:before_updated)
-                  child_block.update
-                  child.else_active = 1
+                  
+                  Hokusai.update(child_block)
+                  child.else_active = true
                 end
               end
             end